move Statistics.scala from package util to package utils
